move auth GVK typedefs to subdirectory
