+++ api/auth/v1alpha1/virtualkey_types.go
