+++ internal/controller/virtualkey_controller_test.go
@@ -28,7 +28,7 @@ import (
 
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
-	authv1alpha1 "github.com/bbdsoftware/litellm-operator/api/v1alpha1"
+	authv1alpha1 "github.com/bbdsoftware/litellm-operator/api/auth/v1alpha1"
 	"github.com/bbdsoftware/litellm-operator/internal/litellm"
 )
